pyproject.toml
[project]
name = "onchain-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agno>=2.2.3",
    "google-genai>=1.48.0",
    "openai>=2.6.1",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.44",
    "stellar-sdk>=13.1.0",
    "streamlit>=1.51.0",
]
